Move GPS configuration to GPS tab (#3326)
