+++ src/css/tabs/configuration.less
@@ -111,40 +111,6 @@
 			margin-bottom: 0;
 		}
 	}
-	.gps {
-		.line {
-			clear: left;
-		}
-		select {
-			float: left;
-			width: 100px;
-			height: 20px;
-			margin: 0 10px 5px 0;
-			border: 1px solid var(--subtleAccent);
-		}
-		.select {
-			>div {
-				&:first-child {
-					float: left;
-					height: 20px;
-					margin-right: 15px;
-					margin-left: 3px;
-				}
-			}
-		}
-		span {
-			line-height: 20px;
-		}
-		.gui_box {
-			float: left;
-			margin-bottom: 10px;
-		}
-		td {
-			&:nth-child(2) {
-				width: 38px;
-			}
-		}
-	}
 	.freelabel {
 		margin-left: 10px;
 		position: relative;
@@ -339,11 +305,6 @@
 		margin-top: 10px;
 		float: left;
 		width: 100%;
-	}
-	.gpsSettings {
-		.note {
-			margin-top: 10px;
-		}
 	}
 }
 @media only screen and (max-width: 1055px) {
