+++ src/js/tabs/gps.js
@@ -1,22 +1,31 @@
 import { i18n } from "../localization";
+import semver from 'semver';
+import { API_VERSION_1_43 } from '../data_storage';
 import GUI, { TABS } from '../gui';
-import { have_sensor } from "../sensor_helpers";
 import FC from '../fc';
 import MSP from "../msp";
 import MSPCodes from "../msp/MSPCodes";
+import { gui_log } from '../gui_log';
+import { have_sensor } from "../sensor_helpers";
+import { mspHelper } from '../msp/MSPHelper';
+import { reinitializeConnection } from '../serial_backend';
+import { updateTabList } from '../utils/updateTabList';
 
 const gps = {};
-gps.initialize = function (callback) {
 
-    if (GUI.active_tab !== 'gps') {
-        GUI.active_tab = 'gps';
-    }
+gps.initialize = async function (callback) {
+
+    GUI.active_tab = 'gps';
+
+    await MSP.promise(MSPCodes.MSP_FEATURE_CONFIG);
+    await MSP.promise(MSPCodes.MSP_GPS_CONFIG);
+    await MSP.promise(MSPCodes.MSP_STATUS);
+
+    load_html();
 
     function load_html() {
         $('#content').load("./tabs/gps.html", process_html);
     }
-
-    MSP.send_message(MSPCodes.MSP_STATUS, false, false, load_html);
 
     function set_online(){
         $('#connect').hide();
@@ -49,6 +58,122 @@ gps.initialize = function (callback) {
         // To not flicker the divs while the fix is unstable
         let gpsWasFixed = false;
 
+        // GPS Configuration
+        const features_e = $('.tab-gps .features');
+
+        FC.FEATURE_CONFIG.features.generateElements(features_e);
+
+        const checkUpdateGpsControls = () => $('.gpsSettings').toggle(FC.FEATURE_CONFIG.features.isEnabled('GPS'));
+
+        $('input.feature', features_e).on('change', function () {
+            const element = $(this);
+
+            FC.FEATURE_CONFIG.features.updateData(element);
+            updateTabList(FC.FEATURE_CONFIG.features);
+
+            if (element.attr('name') === 'GPS') {
+                checkUpdateGpsControls();
+            }
+        });
+
+        checkUpdateGpsControls();
+
+        // generate GPS
+        const gpsProtocols = [
+            'NMEA',
+            'UBLOX',
+            'MSP',
+        ];
+
+        const gpsBaudRates = [
+            '115200',
+            '57600',
+            '38400',
+            '19200',
+            '9600',
+        ];
+
+        const gpsSbas = [
+            i18n.getMessage('gpsSbasAutoDetect'),
+            i18n.getMessage('gpsSbasEuropeanEGNOS'),
+            i18n.getMessage('gpsSbasNorthAmericanWAAS'),
+            i18n.getMessage('gpsSbasJapaneseMSAS'),
+            i18n.getMessage('gpsSbasIndianGAGAN'),
+        ];
+
+        if (semver.gte(FC.CONFIG.apiVersion, API_VERSION_1_43)) {
+            gpsSbas.push(i18n.getMessage('gpsSbasNone'));
+        }
+
+        const gpsProtocolElement = $('select.gps_protocol');
+        const gpsAutoBaudElement = $('input[name="gps_auto_baud"]');
+        const gpsAutoBaudGroup = $('.gps_auto_baud');
+        const gpsAutoConfigElement = $('input[name="gps_auto_config"]');
+        const gpsAutoConfigGroup = $('.gps_auto_config');
+        const gpsUbloxGalileoElement = $('input[name="gps_ublox_galileo"]');
+        const gpsUbloxGalileoGroup = $('.gps_ublox_galileo');
+        const gpsUbloxSbasElement = $('select.gps_ubx_sbas');
+        const gpsUbloxSbasGroup = $('.gps_ubx_sbas');
+        const gpsHomeOnceElement = $('input[name="gps_home_once"]');
+        const gpsBaudrateElement = $('select.gps_baudrate');
+
+        for (let protocolIndex = 0; protocolIndex < gpsProtocols.length; protocolIndex++) {
+            gpsProtocolElement.append(`<option value="${protocolIndex}">${gpsProtocols[protocolIndex]}</option>`);
+        }
+
+        gpsProtocolElement.change(function () {
+            FC.GPS_CONFIG.provider = parseInt($(this).val());
+
+            // Call this to enable or disable auto config elements depending on the protocol
+            gpsAutoConfigElement.change();
+
+        }).val(FC.GPS_CONFIG.provider).change();
+
+        gpsAutoBaudElement.prop('checked', FC.GPS_CONFIG.auto_baud === 1);
+
+        gpsAutoConfigElement.on('change', function () {
+            const checked = $(this).is(":checked");
+
+            const ubloxSelected = FC.GPS_CONFIG.provider === gpsProtocols.indexOf('UBLOX');
+            const mspSelected = FC.GPS_CONFIG.provider === gpsProtocols.indexOf('MSP');
+
+            const enableGalileoVisible = checked && ubloxSelected && semver.gte(FC.CONFIG.apiVersion, API_VERSION_1_43);
+            gpsUbloxGalileoGroup.toggle(enableGalileoVisible);
+
+            const enableSbasVisible = checked && ubloxSelected;
+            gpsUbloxSbasGroup.toggle(enableSbasVisible);
+
+            gpsAutoBaudGroup.toggle(ubloxSelected || mspSelected);
+            gpsAutoConfigGroup.toggle(ubloxSelected || mspSelected);
+
+        }).prop('checked', FC.GPS_CONFIG.auto_config === 1).trigger('change');
+
+        gpsUbloxGalileoElement.change(function() {
+            FC.GPS_CONFIG.ublox_use_galileo = $(this).is(':checked') ? 1 : 0;
+        }).prop('checked', FC.GPS_CONFIG.ublox_use_galileo > 0).change();
+
+        for (let sbasIndex = 0; sbasIndex < gpsSbas.length; sbasIndex++) {
+            gpsUbloxSbasElement.append(`<option value="${sbasIndex}">${gpsSbas[sbasIndex]}</option>`);
+        }
+
+        gpsUbloxSbasElement.change(function () {
+            FC.GPS_CONFIG.ublox_sbas = parseInt($(this).val());
+        }).val(FC.GPS_CONFIG.ublox_sbas);
+
+        $('.gps_home_once').toggle(semver.gte(FC.CONFIG.apiVersion, API_VERSION_1_43));
+        gpsHomeOnceElement.change(function() {
+            FC.GPS_CONFIG.home_point_once = $(this).is(':checked') ? 1 : 0;
+        }).prop('checked', FC.GPS_CONFIG.home_point_once > 0).change();
+
+        for (const baudRate of gpsBaudRates) {
+            gpsBaudrateElement.append(`<option value="${baudRate}">${baudRate}</option>`);
+        }
+
+        gpsBaudrateElement.prop("disabled", true);
+        gpsBaudrateElement.parent().hide();
+
+        // End GPS Configuration
+
         function update_ui() {
             const lat = FC.GPS_DATA.lat / 10000000;
             const lon = FC.GPS_DATA.lon / 10000000;
@@ -80,6 +205,7 @@ gps.initialize = function (callback) {
                     <td style="width: 53%;" i18n="gpsSignalStatusQly">${i18n.getMessage('gpsSignalStatusQly')}</td>
                 </tr>
             `);
+
             if (FC.GPS_DATA.chn.length <= 16) {
                 // Legacy code path: old BF firmware or old ublox module
                 for (let i = 0; i < FC.GPS_DATA.chn.length; i++) {
@@ -155,11 +281,9 @@ gps.initialize = function (callback) {
                     message.action = 'nofix';
                     frame.contentWindow.postMessage(message, '*');
                 }
-            }else{
+            } else {
                 gpsWasFixed = false;
-                $('#connect').show();
-                $('#waiting').hide();
-                $('#loadmap').hide();
+                set_offline();
             }
         }
 
@@ -216,9 +340,31 @@ gps.initialize = function (callback) {
             frame.contentWindow.postMessage(message, '*');
         });
 
+        $('a.save').on('click', function() {
+            // fill some data
+            FC.GPS_CONFIG.auto_baud = $('input[name="gps_auto_baud"]').is(':checked') ? 1 : 0;
+            FC.GPS_CONFIG.auto_config = $('input[name="gps_auto_config"]').is(':checked') ? 1 : 0;
+
+            async function saveConfiguration() {
+                await MSP.promise(MSPCodes.MSP_SET_FEATURE_CONFIG, mspHelper.crunch(MSPCodes.MSP_SET_FEATURE_CONFIG));
+                await MSP.promise(MSPCodes.MSP_SET_GPS_CONFIG, mspHelper.crunch(MSPCodes.MSP_SET_GPS_CONFIG));
+                await MSP.promise(MSPCodes.MSP_EEPROM_WRITE);
+                reboot();
+            }
+
+            function reboot() {
+                gui_log(i18n.getMessage('configurationEepromSaved'));
+
+                GUI.tab_switch_cleanup(function() {
+                    MSP.send_message(MSPCodes.MSP_SET_REBOOT, false, false, reinitializeConnection);
+                });
+            }
+
+            saveConfiguration();
+        });
+
         GUI.content_ready(callback);
     }
-
 };
 
 gps.cleanup = function (callback) {
